fix long to int

+++ api/@ohos.multimedia.systemSoundManager.d.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023 Huawei Device Co., Ltd.
+ * Copyright (c) 2023-2025 Huawei Device Co., Ltd.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -261,39 +261,43 @@ declare namespace systemSoundManager {
 
   /**
    * Define the ringtone category.
+   * @constant
    * @syscap SystemCapability.Multimedia.SystemSound.Core
    * @systemapi
    * @since 12 dynamic
    * @since 22 static
    */
-  const TONE_CATEGORY_RINGTONE: long;
+  const TONE_CATEGORY_RINGTONE: int;
 
   /**
    * Define the text message tone category.
+   * @constant
    * @syscap SystemCapability.Multimedia.SystemSound.Core
    * @systemapi
    * @since 12 dynamic
    * @since 22 static
    */
-  const TONE_CATEGORY_TEXT_MESSAGE: long;
+  const TONE_CATEGORY_TEXT_MESSAGE: int;
 
   /**
    * Define the notification tone category.
+   * @constant
    * @syscap SystemCapability.Multimedia.SystemSound.Core
    * @systemapi
    * @since 12 dynamic
    * @since 22 static
    */
-  const TONE_CATEGORY_NOTIFICATION: long;
+  const TONE_CATEGORY_NOTIFICATION: int;
 
   /**
    * Define the alarm tone category.
+   * @constant
    * @syscap SystemCapability.Multimedia.SystemSound.Core
    * @systemapi
    * @since 12 dynamic
    * @since 22 static
    */
-  const TONE_CATEGORY_ALARM: long;
+  const TONE_CATEGORY_ALARM: int;
 
   /**
    * Define the contact tone category.
@@ -303,7 +307,7 @@ declare namespace systemSoundManager {
    * @since 20 dynamic
    * @since 22 static
    */
-  const TONE_CATEGORY_CONTACTS: long;
+  const TONE_CATEGORY_CONTACTS: int;
 
   /**
    * Define the app notification tone category.
@@ -398,7 +402,7 @@ declare namespace systemSoundManager {
 
     /**
      * Sets tone category.
-     * @param { long } category - tone category. This parameter can be one of {@link TONE_CATEGORY_RINGTONE},
+     * @param { int } category - tone category. This parameter can be one of {@link TONE_CATEGORY_RINGTONE},
      * {@link TONE_CATEGORY_TEXT_MESSAGE}, {@link TONE_CATEGORY_NOTIFICATION}, {@link TONE_CATEGORY_ALARM}.
      * In addition, this parameter can be result of OR logical operator of these constants.
      * @throws { BusinessError } 202 - Caller is not a system application.
@@ -410,11 +414,11 @@ declare namespace systemSoundManager {
      * @since 12 dynamic
      * @since 22 static
      */
-    setCategory(category: long): void;
+    setCategory(category: int): void;
 
     /**
      * Gets tone category.
-     * @returns { long } Tone category. This value can be one of {@link TONE_CATEGORY_RINGTONE},
+     * @returns { int } Tone category. This value can be one of {@link TONE_CATEGORY_RINGTONE},
      * {@link TONE_CATEGORY_TEXT_MESSAGE}, {@link TONE_CATEGORY_NOTIFICATION}, {@link TONE_CATEGORY_ALARM}.
      * In addition, this value can be result of OR logical operator of these constants.
      * @throws { BusinessError } 202 - Caller is not a system application.
@@ -423,7 +427,7 @@ declare namespace systemSoundManager {
      * @since 12 dynamic
      * @since 22 static
      */
-    getCategory(): long;
+    getCategory(): int;
 
     /**
      * Sets media type.
